--- MotyokoiApplication/app/src/main/java/com/example/motyokoiapplication/ListViewActivity.java
@@ -1,6 +1,0 @@
-package com.example.motyokoiapplication;
-
-import android.app.Activity;
-
-public class ListViewActivity extends Activity {
-}
