+++ MotyokoiApplication/app/src/main/java/com/example/motyokoiapplication/ui/dashboard/DashboardViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.motyokoiapplication.ui.dashboard
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class DashboardViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
-    }
-    val text: LiveData<String> = _text
-}
+//package com.example.motyokoiapplication.ui.dashboard
+//
+//import androidx.lifecycle.LiveData
+//import androidx.lifecycle.MutableLiveData
+//import androidx.lifecycle.ViewModel
+//
+//class DashboardViewModel : ViewModel() {
+//
+//    private val _text = MutableLiveData<String>().apply {
+//        value = "This is dashboard Fragment"
+//    }
+//    val text: LiveData<String> = _text
+//}
