+++ MotyokoiApplication/app/src/main/java/com/example/motyokoiapplication/ui/dashboard/User.kt
@@ -1,0 +1,3 @@
+package com.example.motyokoiapplication.ui.dashboard
+
+data class User (val name: String, val email: String, val place: String)
