+++ MotyokoiApplication/app/src/main/java/com/example/motyokoiapplication/ui/notifications/User.kt
@@ -1,3 +1,3 @@
-package com.example.motyokoiapplication.ui.notifications
-
-class User (val name: String, val email: String, val place: String)
+//package com.example.motyokoiapplication.ui.notifications
+//
+//class User (val name: String, val email: String, val place: String)
